No Sleep 1.21.0.0
I updated No Sleep for 1.21. It is no longer compatible with old versions, mainly to make it easier for me to maintain. The version format is also different. It now has four numbers. The first three match the intended Minecraft version, and the last one is for updates within the same Minecraft version. I also fixed a bug where phantoms would presumably spawn at the world origin instead of above the player. I do not know if this bug existed before..? But it's fixed now regardless.