begin developing Infoverlays
I think I said somewhere that Minimal Structures was the next data pack, but ADHD. In this commit, I added concepts for two more packs: Elemental Enchanting and Infoverlays. Infoverlays is the one I am starting first. See the updated README.md (which also has an updated introduction and format) for details! I have created a functioning overlay for compasses and filled maps which displays your coordinates and how each would be affected if you moved forward. The way it is checked currently is will not scale. In the future, I will have to create a system that prioritizes the main hand item over the off hand item. I haven't made it work in item frames yet either. That will come in a later update, along with the priority system and probably another overlay such as that of the daylight detector. I also removed the No Sleep ZIP file that I probably included accidentally, and changed the erroneous pack format of No Sleep (15) to the correct value of 48.